rolevate-agent7/requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
python-dotenv==1.1.1
pydantic==2.12.3
openai==2.5.0
boto3==1.40.55
requests==2.32.5
gql==4.0.0
graphql-core==3.2.6
langgraph>=0.1.0
langchain>=0.1.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
pdfplumber==0.11.7
PyPDF2==3.0.1
python-docx==1.2.0
pytesseract==0.3.13
pillow==12.0.0
beautifulsoup4==4.12.3
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
